move pwa to new lib layout
