+++ projects/storefrontlib/src/cms-structure/index.ts
@@ -1,3 +1,4 @@
 export * from './outlet/index';
 export * from './page/index';
+export * from './pwa/index';
 export * from './seo/index';
