+++ projects/storefrontlib/src/cms-structure/pwa/components/add-to-home-screen-banner/add-to-home-screen-banner.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { AddToHomeScreenService } from '../../services/add-to-home-screen.service';
-import { AddToHomeScreenComponent } from './../add-to-home-screen.component';
+import { AddToHomeScreenComponent } from '../add-to-home-screen.component';
 
 @Component({
   selector: 'cx-add-to-home-screen-banner',
