+++ projects/storefrontlib/src/cms-structure/pwa/components/add-to-home-screen.component.ts
@@ -1,12 +1,12 @@
 import { OnInit } from '@angular/core';
-import { AddToHomeScreenService } from './../services/add-to-home-screen.service';
 import { Observable } from 'rxjs';
+import { AddToHomeScreenService } from '../services/add-to-home-screen.service';
 
 export abstract class AddToHomeScreenComponent implements OnInit {
   canPrompt$: Observable<boolean>;
   constructor(protected addToHomeScreenService: AddToHomeScreenService) {}
 
-  ngOnInit() {
+  ngOnInit(): void {
     this.canPrompt$ = this.addToHomeScreenService.canPrompt$;
   }
