+++ projects/storefrontlib/src/cms-structure/pwa/index.ts
@@ -1,0 +1,5 @@
+export * from './components/add-to-home-screen-banner/add-to-home-screen-banner.component';
+export * from './components/add-to-home-screen-btn/add-to-home-screen-btn.component';
+export * from './components/add-to-home-screen.component';
+export * from './pwa.module';
+export * from './pwa.module-config';
