+++ projects/storefrontlib/src/cms-structure/pwa/pwa.module.ts
@@ -1,15 +1,13 @@
-import { NgModule, APP_INITIALIZER } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { APP_INITIALIZER, NgModule } from '@angular/core';
 import {
   ServiceWorkerModule,
   ɵangular_packages_service_worker_service_worker_b as RegistrationOptions,
 } from '@angular/service-worker';
-
 import { Config, ConfigModule, I18nModule } from '@spartacus/core';
-
-import { defaultPWAModuleConfig, PWAModuleConfig } from './pwa.module-config';
-import { AddToHomeScreenBtnComponent } from './components/add-to-home-screen-btn/add-to-home-screen-btn.component';
 import { AddToHomeScreenBannerComponent } from './components/add-to-home-screen-banner/add-to-home-screen-banner.component';
+import { AddToHomeScreenBtnComponent } from './components/add-to-home-screen-btn/add-to-home-screen-btn.component';
+import { defaultPWAModuleConfig, PWAModuleConfig } from './pwa.module-config';
 import { AddToHomeScreenService } from './services/add-to-home-screen.service';
 
 export function pwaConfigurationFactory(
@@ -18,7 +16,7 @@ export function pwaConfigurationFactory(
   return { enabled: (pwaConfig.production && pwaConfig.pwa.enabled) || false };
 }
 
-export function pwaFactory(addToHomeScreenService) {
+export function pwaFactory(addToHomeScreenService): any {
   const result = () => addToHomeScreenService;
   return result;
 }
